cmd/compile/internal/devirtualize: update PGO devirt test to include a method from another package; devirt.pprof not yet updated
